Move filter predicate class under `EventApplier`.
