+++ server/src/main/java/org/spine3/server/aggregate/EventApplier.java
@@ -20,6 +20,7 @@
 
 package org.spine3.server.aggregate;
 
+import com.google.common.base.Predicate;
 import com.google.protobuf.Message;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -27,17 +28,20 @@ import org.spine3.server.internal.MessageHandlerMethod;
 import org.spine3.server.reflect.MethodMap;
 import org.spine3.server.reflect.Methods;
 
+import javax.annotation.Nullable;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
 import java.util.Map;
+
+import static com.google.common.base.Preconditions.checkNotNull;
 
 /**
  * A wrapper for event applier method.
  *
  * @author Alexander Yevsyukov
  */
-class EventApplier extends MessageHandlerMethod<Aggregate, Void> {
+/* package */ class EventApplier extends MessageHandlerMethod<Aggregate, Void> {
 
     /**
      * Creates a new instance to wrap {@code method} on {@code target}.
@@ -84,4 +88,41 @@ class EventApplier extends MessageHandlerMethod<Aggregate, Void> {
         return LogSingleton.INSTANCE.value;
     }
 
+    /**
+     * The predicate for filtering event applier methods.
+     */
+    /* package */ static class FilterPredicate implements Predicate<Method> {
+
+        private static final int EVENT_PARAM_INDEX = 0;
+
+        /**
+         * Checks if a method is an event applier.
+         *
+         * @param method to check
+         * @return {@code true} if the method is an event applier, {@code false} otherwise
+         */
+        public static boolean isEventApplier(Method method) {
+            final boolean isAnnotated = method.isAnnotationPresent(Apply.class);
+            if (!isAnnotated) {
+                return false;
+            }
+
+            final Class<?>[] parameterTypes = method.getParameterTypes();
+            final boolean hasOneParam = parameterTypes.length == 1;
+            if (!hasOneParam) {
+                return false;
+            }
+
+            final boolean firstParamIsMessage = Message.class.isAssignableFrom(parameterTypes[EVENT_PARAM_INDEX]);
+            final boolean returnsNothing = Void.TYPE.equals(method.getReturnType());
+
+            return firstParamIsMessage && returnsNothing;
+        }
+
+        @Override
+        public boolean apply(@Nullable Method method) {
+            checkNotNull(method);
+            return isEventApplier(method);
+        }
+    }
 }
